requirements.txt
# Professional Research Skill - Python Dependencies
# Install: python -m pip install -r requirements.txt

# Core (required)
google-generativeai>=0.8.0
python-docx>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0

# Financial data
akshare>=1.12.0
yfinance>=0.2.30

# PDF parsing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# NAS full-text search index
whoosh>=2.7.4

# Academic / RSS
feedparser>=6.0.0

# Optional (charts & data processing)
# matplotlib>=3.8.0
# pandas>=2.1.0
# seaborn>=0.13.0
